refac (Line, SectionAcceptanceTest) : 구간 관련 인수테스트를 구간 인수테스트 클래스로 이동 리팩토링
